third_party/blink/web_tests/media/mediasession/mojo/set-null-metadata.html

<!DOCTYPE html>
<title>MediaSession Mojo Test</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/mediasession/media_session.mojom.js"></script>
<script src="resources/mediasessionservice-mock.js"></script>
<script src="resources/utils.js"></script>
<script>

async_test(function(t) {
  let m = mediaSessionServiceMock;
  m.setMetadataCallback(t.step_func(function(receivedMetadata) {
    assert_equals(receivedMetadata, null);
    t.done();
  }));
  window.navigator.mediaSession.metadata = null;
}, "test that null MediaMetadata is correctly propagated");

</script>
